+++ src/OpenTelemetry/CHANGELOG.md
@@ -36,6 +36,8 @@
   [#1135](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1135))
 * Renamed `ParentOrElseSampler` to `ParentBasedSampler`
   ([#1173](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1173))
+* Renamed `ProbabilitySampler` to `TraceIdRatioBasedSampler`
+  ([#1174](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1174))
 
 ## 0.4.0-beta.2